chore: merge customers and use-cases into one view
