+++ src/components/patterns/Footer/Footer.js
@@ -64,7 +64,6 @@ const FOOTER_COLUMNS = [
     title: 'Company',
     links: [
       { label: 'About', href: '/about' },
-      { label: 'Customers', href: '/customers' },
       { label: 'Bug Reports', href: issueUrl.bug() },
       { label: 'Contact', href: 'mailto:hello@microlink.io' },
       { label: 'GitHub', href: 'https://github.com/microlinkhq' },
