+++ src/pages/embed/index.js
@@ -2011,7 +2011,7 @@ const CustomerStories = () => {
         {stories.map(({ slug, name, blurb, icon }) => (
           <CustomerStoryCard
             key={slug}
-            href={`/customers/${slug}`}
+            href={`/use-cases/${slug}`}
             data-event-location='Embed'
             data-event-name={`Customer Story ${name}`}
           >
