+++ src/pages/use-cases/luckynote.js
